+++ unittest/std_map.cpp
@@ -3,7 +3,7 @@
 
 #include <eigenpy/eigenpy.hpp>
 #include <eigenpy/std-map.hpp>
-#include <iostream>
+#include <boost/unordered_map.hpp>
 
 namespace bp = boost::python;
 
@@ -22,14 +22,32 @@ std::map<std::string, T1> copy(const std::map<std::string, T1>& map) {
   return out;
 }
 
+template <typename T1>
+boost::unordered_map<std::string, T1> copy_boost(
+    const boost::unordered_map<std::string, T1>& obj) {
+  return obj;
+}
+
+struct X {
+  X() = delete;
+  X(int x) : val(x) {}
+  int val;
+};
+
 BOOST_PYTHON_MODULE(std_map) {
   eigenpy::enableEigenPy();
 
-  eigenpy::python::StdMapPythonVisitor<
+  eigenpy::StdMapPythonVisitor<
       std::string, double, std::less<std::string>,
       std::allocator<std::pair<const std::string, double> >,
       true>::expose("StdMap_Double");
 
+  eigenpy::GenericMapVisitor<boost::unordered_map<std::string, int> >::expose(
+      "boost_map_int");
+
+  eigenpy::GenericMapVisitor<std::map<std::string, X> >::expose("StdMap_X");
+
   bp::def("std_map_to_dict", std_map_to_dict<double>);
   bp::def("copy", copy<double>);
+  bp::def("copy_boost", copy_boost<int>);
 }